De-mongo: clean up migration code and mongo database connection.
