--- db/mongo.go
@@ -1,143 +1,0 @@
-package db
-
-// Wraps an mgo connection and db object for convenience
-// Yes, these are globals. I'm undecided, but let's see how it goes.
-
-import (
-	"errors"
-	"fmt"
-	"strings"
-	"sync"
-	"time"
-
-	"github.com/fluffle/golog/logging"
-	"gopkg.in/mgo.v2"
-	"gopkg.in/mgo.v2/bson"
-)
-
-const DATABASE string = "sp0rkle"
-
-type mongoDatabase struct {
-	sync.Mutex
-	sessions []*mgo.Session
-	live     bool
-}
-
-var Mongo = &mongoDatabase{}
-
-func (m *mongoDatabase) Init(db string) error {
-	m.Lock()
-	defer m.Unlock()
-	if m.sessions != nil {
-		return errors.New("init already called")
-	}
-	s, err := mgo.Dial(db)
-	if err != nil {
-		return err
-	}
-	// Let's be explicit about requiring journaling, ehh?
-	s.EnsureSafe(&mgo.Safe{J: true})
-	// The markov db is large and we get timeouts querying it
-	// on the shitty f1 micro sp0rkle lives on.
-	s.SetSocketTimeout(10 * time.Minute)
-	s.SetSyncTimeout(10 * time.Minute)
-	m.sessions = []*mgo.Session{s}
-	m.live = true
-	return nil
-}
-
-func (m *mongoDatabase) Live() bool { return m.live }
-
-func (m *mongoDatabase) Close() {
-	m.Lock()
-	defer m.Unlock()
-	if m.sessions == nil {
-		return
-	}
-	for _, s := range m.sessions {
-		s.Close()
-	}
-	m.sessions = nil
-	m.live = false
-}
-
-func (m *mongoDatabase) C(name string) Collection {
-	m.Lock()
-	defer m.Unlock()
-	if !m.live || m.sessions == nil {
-		logging.Fatal("Tried to create MongoDB collection %q when disconnected.", name)
-	}
-	s := m.sessions[0].Copy()
-	s.SetSocketTimeout(10 * time.Minute)
-	s.SetSyncTimeout(10 * time.Minute)
-	m.sessions = append(m.sessions, s)
-	return &mongoCollection{Collection: s.DB(DATABASE).C(name)}
-}
-
-type mongoCollection struct {
-	*mgo.Collection
-	debug_ bool
-}
-
-func (m *mongoCollection) Debug(on bool) {
-	m.debug_ = on
-}
-
-func (m *mongoCollection) debug(f string, args ...any) {
-	if m.debug_ {
-		logging.Debug("MONGO: "+f, args...)
-	}
-}
-
-func (m *mongoCollection) Get(key Key, value any) error {
-	k := key.M()
-	m.debug("Get(%v)", k)
-	return m.Collection.Find(k).One(value)
-}
-
-func (m *mongoCollection) Match(key, regex string, value any) error {
-	q := bson.M{strings.ToLower(key): bson.M{"$regex": regex, "$options": "i"}}
-	return m.Collection.Find(q).All(value)
-}
-
-func (m *mongoCollection) All(key Key, value any) error {
-	return m.Collection.Find(key.M()).All(value)
-}
-
-func (m *mongoCollection) Fsck(value any) error {
-	panic("srsly dude")
-}
-
-func (m *mongoCollection) Put(value any) (err error) {
-	switch value := value.(type) {
-	case Keyer:
-		_, err = m.Collection.Upsert(value.K().M(), value)
-	case Indexer:
-		_, err = m.Collection.UpsertId(value.Id(), value)
-	default:
-		return fmt.Errorf("put: don't know how to put value %#v", value)
-	}
-	return err
-}
-
-func (m *mongoCollection) BatchPut(value any) error {
-	panic("no batch puts for you")
-}
-
-func (m *mongoCollection) Del(value any) error {
-	switch value := value.(type) {
-	case Keyer:
-		return m.Collection.Remove(value.K().M())
-	case Indexer:
-		return m.Collection.RemoveId(value.Id())
-	}
-	return fmt.Errorf("del: don't know how to delete value %#v", value)
-}
-
-func (m *mongoCollection) Next(k Key, set ...int) (int, error) {
-	panic("no autoincrements for you")
-}
-
-func (m *mongoCollection) Mongo() *mgo.Collection {
-	return m.Collection
-}
